% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APAhtmlTable.R
\name{APAhtmlTable}
\alias{APAhtmlTable}
\title{A Function to Create APA Formatted HTML Tables}
\usage{
APAhtmlTable(df = NULL, round = TRUE, tidy = TRUE, digits = 3,
  genNote = NULL, font = "Times New Roman", fontSize = 12,
  tableNum = NULL, tableTitle = NULL, headerBold = FALSE,
  headerAlign = "center", leftColAlign = "left", contentAlign = "center",
  header = NULL)
}
\arguments{
\item{df}{Data Frame or Function output to print in an APA formatted HTML table. Defaults to NULL.}

\item{tidy}{Do you want to convert the object into a data frame (TRUE/FALSE). This is accomplished using the Tidy() function in the package Broom. Defaults to TRUE.}

\item{digits}{Number of digits you want to round numbers too. Leave blank if you don't want to round numbers or if the Data Frame contains strings. Defaults to NULL.}

\item{genNote}{General note to be displayed below the table (Automatically begins with NOTE. "Your note here"). Defaults to NULL.}

\item{font}{Sets the font that text is displayed in. Defaults to Times New Roman.}

\item{tableNum}{Adds a table number above the table. Format is "Table XX". If NULL, no table number is displayed. Defualts to NULL.}

\item{tableTitle}{Adds a title above the table. If NULL, no title will be displayed. Defaults to NULL.}

\item{headerBold}{Do you want the column headers to be bolded (TRUE/FALSE)? Defaults to FALSE.}

\item{headerAlign}{Sets the horizontal alignment of the column headers ("left","center","right"). Defaults to "center".}

\item{leftColAlign}{Sets the horizontal alignment of the first column in the table ("left","center","right"). Defaults to "left".}

\item{contentAlign}{Sets the horizontal alignment of all other cells in the table besides the header (see headerAlign) and left column (see leftColAlign).}

\item{header}{Manually define the column headers as a list. The length of the list must match the number of columns in the table. Defaults to NULL.}

\item{fontsize}{Sets the font size of all text in the table. Defaults to 12pt.}
}
\description{
This function allows you to print the output of data frames and most R functions in an APA formatted table.
}
\examples{
lm.obj <- lm(mpg~am+disp, data=mtcars)
lm.summary <- summary(lm.obj)
APAhtmlTable(df=lm.summary, digits=3, tableNum=1, tableTitle="Miles per Gallon Regressed on Transmission Type and Displacement", header=c("Term","Estimate","SE","t","p"))
}
\keyword{APA}
\keyword{Tables}
