% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormatrix.R
\name{cormatrix}
\alias{cormatrix}
\title{A Function to Create a Correlation Matrix}
\usage{
cormatrix(ivs = NULL, data = "NULL", method = "pearson", output = "r",
  means = FALSE, columnnames = TRUE, CI = FALSE, CIlevel = 0.95,
  displayDF = FALSE, displayP = FALSE, showSig = TRUE, digits = 3)
}
\arguments{
\item{ivs}{List of variables from which to compute all pairwise correlations. Input must be a list (i.e., c("var1","var2")). Defaults to NULL.}

\item{data}{A data frame containing the observed variables to correlate. Defaults to NULL.}

\item{method}{Type of correlation coefficient to computed. Options include "pearson","kendall", or "spearman". Defaults to "pearson".}

\item{output}{the format desired for output. Options include "html" or "r". Defaults to "r".}

\item{columnnames}{Option to determine whether variable names are used as column headers (TRUE) or numbers are used instead (FALSE). Defaults to TRUE.}

\item{CI}{Option to include confidence intervals in the correlation matrix table. Defaults to FALSE.}

\item{CIlevel}{If confidence intervals are requested, this sets the desired confidence interval. Defaults to .95}

\item{displayDF}{Option to include the degrees of freedom used in each test in the correlation matrix table. Defults to FALSE.}

\item{displayP}{Option to include the statistical significance (p-value) of each correlation in the matrix. Null hypothesis that there is no relationship between the variables (r=0). Defaults to FALSE.}

\item{showSig}{Option to include symbols beside correlation coefficient denoting the level of significance of each correlation. Levels include p < .1, p < .05, p < .1, p < .001. Defaults to FALSE.}

\item{digits}{Number of digits you want to round numbers too. Leave blank if you don't want to round numbers or if the Data Frame contains strings. Defaults to NULL.}
}
\description{
This function allows you to print the output of data frames and most R functions in an APA formatted table.
}
\examples{
cormatrix(ivs = c("mpg","wt","hp"), data=mtcars, digits=3, output="html", columnnames = FALSE, CI=TRUE, CIlevel=.95, displayDF=TRUE, displayP=TRUE, showSig=TRUE)
}
\keyword{Correlation}
