% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df.R
\name{round_df}
\alias{round_df}
\title{A Function that rounds all numeric values in a dataframe to the specifed number of decimal places.}
\usage{
round_df(df = NULL, digits = 3)
}
\arguments{
\item{df}{Data Frame you want to round. Defaults to NULL.}

\item{digits}{Number of digits you want to round numbers too. Leave blank if you don't want to round numbers or if the Data Frame contains strings. Defaults to 3.}
}
\description{
This function rounds all numeric values in a dataframe to the specifed number of decimal places.
}
\examples{
APAhtmlTable()
}
\keyword{round}
