% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniANOVA.R
\name{uniANOVA}
\alias{uniANOVA}
\title{A Function to Conduct Univariate Analysis of (Co)Variance.}
\usage{
uniANOVA(dv = NULL, formula = NULL, factors = NULL, cov = NULL,
  data = NULL, descriptives = TRUE, assumptions = FALSE, ss = 3,
  digits = 3, emmeans = FALSE, conf = 0.95, comparisons = NULL,
  postCorr = "none", plotBar = FALSE, xLabel = NULL, yLabel = NULL,
  fillLabel = NULL, yLower = NULL, yUpper = NULL, yBreaks = NULL,
  output = "r")
}
\arguments{
\item{dv}{Dependent variable. Defaults to NULL.}

\item{factors}{Independent variables (converted to factors). All factors will be multiplied in a full-factorial design. Any factors that you do not wish to be crossed with the others should be entere using the "cov=" command. Defaults to NULL.}

\item{cov}{Independent variables and/or covariates not to be crossed with other factors. These variables will also not be used for grouping in descriptive statistics. You may still utilize these factors in simple and main effects tests using the "comparisons=" command. Defaults to NULL.}

\item{data}{A data frame containing the observed variables to correlate. Defaults to NULL.}

\item{descriptives}{TRUE/FALSE indicating whether to display descriptive statistics for the dependent variable grouped by factors. Defaults to TRUE.}

\item{assumptions}{TRUE/FALSE indicating whether you would like to print tests of model assumptions and plots of residuals.}

\item{ss}{A number indicating which type sum of squares to use (i.e., 1,2,3). Defaults to 3. Uses the CAR package to conduct ANOVAs with type 3 sums of squares.}

\item{digits}{Number of digits you want to round numbers too. Leave blank if you don't want to round numbers or if the Data Frame contains strings. Defaults to NULL.}

\item{emmeans}{TRUE/FALSE indicating whether to display estimated marginal means grouped by factors.}

\item{conf}{A number indicating the desired confidence interval. Defaults to .95.}

\item{comparisons}{A list of comparisons for which you would like to conduct main or simple effects tests. Prints estimated marginal means and tests of effects (including effect size) for each comparison specified in the list. Defaults to NULL.}

\item{plotBar}{TRUE/FALSE indicating whether a QQPlot bar plot should be returned for estimated marginal means across all factors, with error bars indicating confidence intervals at levels set by "conf". Defaults to FALSE.}

\item{xLabel}{A string indicating the label for the x-axis in the bar plot if one is requested. Defaults to NULL, which will display variable name.}

\item{yLabel}{A string indicating the label for the y-axis in the bar plot if one is requested. Defaults to NULL, which will display variable name.}

\item{fillLabel}{A string indicating the label for the fill variable in the bar plot if one is requested. Defaults to NULL, which will display variable name.}

\item{yLower}{A number indicating the lower limit of the y-axis in the bar plot if requested. Defaults to NULL, in which case limits are set by QQPlot.}

\item{yUpper}{A number indicating the upper limit of the y-axis in the bar plot if requested. Defaults to NULL, in which case limits are set by QQPlot.}

\item{yBreaks}{A number indicating the break value (i.e., interval of numbers and ticks) for the y-axis in the bar plot if requested. Defaults to NULL, in which case limits are set by QQPlot.}

\item{output}{A string indicating whether results should be displayed as R output (output="r") or HTML output "output="html". Defaults to r output.}

\item{postcor}{A string indicating type of correction to apply to main or simple effects tests requested using the "comparisons=" commond. Options include "none", tukey", "scheffe", "sidak", "bonferroni", "dunnettx", "mvt". Defaults to "none".}
}
\description{
This function conducts a univariate analysis of (co)variance (ANOVA/ANCOVA). It includes options to print descriptives, tests and plots of model assumptions, estimated marginal means, main and simple effects tests, and APA formatted bar plots.
}
\examples{
uniANOVA(dv = "mpg", factors = c("vs","am"), cov=c("hp"), data = mtcars, descriptives=TRUE, assumptions=TRUE, ss=3, digits=3, emmeans=TRUE, conf=.95, comparisons=c("am","vs","vs*am"), plotBar=TRUE, xLabel="Engine", yLabel="Miles per Gallon (US)", fillLabel="Transmission" yLower=10,yUpper=45,yBreaks=5,output = "html").
}
\keyword{ANOVA}
